feat: update scene stub
